+++ src/nl/hannahsten/texifyidea/index/file/LatexIndexableSetContributor.kt
@@ -80,9 +80,7 @@ class LatexIndexableSetContributor : IndexableSetContributor() {
             runInBackgroundBlocking(project, "Searching for inclusions by absolute path...") {
                 // Bibliography and direct input commands
                 val commandNames = CommandMagic.includeOnlyExtensions.entries.filter { it.value.contains("bib") || it.value.contains("tex") }.map { it.key }.toSet()
-                val externalFiles = runReadAction {
-                    LatexIncludesIndex.Util.getCommandsByNames(commandNames, project, GlobalSearchScope.projectScope(project))
-                }
+                val externalFiles = LatexIncludesIndex.Util.getCommandsByNames(commandNames, project, GlobalSearchScope.projectScope(project))
                     // We can't add single files, so take the parent
                     .mapNotNull {
                         val path = runReadAction { it.requiredParameter(0) } ?: return@mapNotNull null