+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Default to 0.15.0 behaviour if inputs field is not found in Tectonic.toml
 
 ### Fixed
+* Fix unicode inspection for \Ohm, by @slideclimb
 * Fix a false negative of the non-matching environment name inspection for empty parameters
 * Fix a possible UI freeze if pdflatex doesn't exit while TeXiFy is checking for files to index
 * Fix StackOverflowError when color definitions reference each other