+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 * Default to 0.15.0 behaviour if inputs field is not found in Tectonic.toml
 
 ### Fixed
-* Fix exceptions #3938, #3939, #3940
+* Fix exceptions #3938, #3939, #3940, #3955
 * Fix a UI freeze which could occur with autocompile on save on project open
 * Fix unicode inspection for \Ohm, by @slideclimb
 * Fix a false negative of the non-matching environment name inspection for empty parameters